refactor(schema): Pull SourceKind out into schemas
