+++ src/cargo/util_schemas/core/mod.rs
@@ -1,0 +1,4 @@
+mod source_kind;
+
+pub use source_kind::GitReference;
+pub use source_kind::SourceKind;
